// ts/features/itwallet/analytics/enum.ts
export enum ITW_ACTIONS_EVENTS {
  CLOSE_BANNER = "CLOSE_BANNER",
  ITW_CLOSE_BANNER = "ITW_CLOSE_BANNER",
  ITW_CREDENTIAL_COPY_LIST_ITEM = "ITW_CREDENTIAL_COPY_LIST_ITEM",
  ITW_CREDENTIAL_QUALIFICATION_DETAIL = "ITW_CREDENTIAL_QUALIFICATION_DETAIL",
  ITW_CREDENTIAL_SHOW_AUTH_SOURCE = "ITW_CREDENTIAL_SHOW_AUTH_SOURCE",
  ITW_CREDENTIAL_SHOW_ISSUER = "ITW_CREDENTIAL_SHOW_ISSUER",
  ITW_CREDENTIAL_START_REISSUING = "ITW_CREDENTIAL_START_REISSUING",
  ITW_DATA_SHARE_ACCEPTED = "ITW_DATA_SHARE_ACCEPTED",
  ITW_ID_METHOD_SELECTED = "ITW_ID_METHOD_SELECTED",
  ITW_KO_STATE_ACTION_SELECTED = "ITW_KO_STATE_ACTION_SELECTED",
  ITW_OPERATION_BLOCK_ACTION = "ITW_OPERATION_BLOCK_ACTION",
  ITW_START_ACTIVATION = "ITW_START_ACTIVATION",
  ITW_START_DEACTIVATION = "ITW_START_DEACTIVATION",
  ITW_TAP_BANNER = "ITW_TAP_BANNER",
  ITW_TOS = "ITW_TOS",
  SURVEY_REQUEST_ACCEPTED = "SURVEY_REQUEST_ACCEPTED",
  SURVEY_REQUEST_DECLINED = "SURVEY_REQUEST_DECLINED",
  TAP_BANNER = "TAP_BANNER",
  WALLET_ADD = "WALLET_ADD",
  WALLET_ADD_START = "WALLET_ADD_START",
  WALLET_CATEGORY_FILTER = "WALLET_CATEGORY_FILTER"
}

export enum ITW_CONFIRM_EVENTS {
  ITW_DEACTIVATED = "ITW_DEACTIVATED",
  ITW_UX_SUCCESS = "ITW_UX_SUCCESS"
}

export enum ITW_ERRORS_EVENTS {
  ITW_ADD_CREDENTIAL_TIMEOUT = "ITW_ADD_CREDENTIAL_TIMEOUT",
  ITW_STATUS_CREDENTIAL_ATTESTATION_FAILURE = "ITW_STATUS_CREDENTIAL_ATTESTATION_FAILURE",
  ITW_TRUSTMARK_RENEW_FAILURE = "ITW_TRUSTMARK_RENEW_FAILURE"
}

export enum ITW_SCREENVIEW_EVENTS {
  BANNER = "BANNER",
  ITW_BANNER = "ITW_BANNER",
  ITW_DATA_SHARE = "ITW_DATA_SHARE",
  ITW_INTRO = "ITW_INTRO",
  ITW_OPERATION_BLOCK = "ITW_OPERATION_BLOCK",
  ITW_SETTINGS = "ITW_SETTINGS",
  SURVEY_REQUEST = "SURVEY_REQUEST",
  WALLET = "WALLET",
  WALLET_ADD_LIST_ITEM = "WALLET_ADD_LIST_ITEM"
}

export enum ITW_TECH_EVENTS {
  ITW_ID_AUTHENTICATION_COMPLETED = "ITW_ID_AUTHENTICATION_COMPLETED",
  ITW_ID_REQUEST = "ITW_ID_REQUEST",
  ITW_ID_VERIFIED_DOCUMENT = "ITW_ID_VERIFIED_DOCUMENT"
}